cleaned up homepage and play a game page
