--- client/.env
@@ -1,1 +1,0 @@
-VITE_CLERK_PUBLISHABLE_KEY=pk_test_Zmx1ZW50LXNhbG1vbi03NC5jbGVyay5hY2NvdW50cy5kZXYk
