+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "client",
+  "name": "aws-game-hackathon-front-end",
   "private": true,
   "version": "0.0.0",
   "type": "module",
