+++ src/Pages/Home.jsx
@@ -2,9 +2,7 @@ import React, { useState, useEffect } from 'react';
 import Navbar from '../CommonComponents/Navbar';
 import HomeTopSection from '../Components/HomeTopSection';
 import WhatGameIs from '../Components/WhatGameIs';
-import Levels from '../Components/Levels';
 import Footer from '../CommonComponents/Footer';
-import Testimonials from '../Components/Testimonials';
 import { useUser } from '@clerk/clerk-react';
 import PopUpCard from '../Components/PopUpCard';
 
@@ -39,12 +37,6 @@ const Home = () => {
       </div>
       <div className="">
         <WhatGameIs />
-      </div>
-      <div className="">
-        <Levels />
-      </div>
-      <div className="">
-        <Testimonials />
       </div>
       <div className="">
         <Footer />
