+++ src/Pages/Levels/Master.json
@@ -1,9 +1,9 @@
 [
   {
     "id": 1,
-    "question": "a__reciation",
-    "options": ["P", "F", "X"],
-    "answer": "P"
+    "question": "_ompli_ated",
+    "options": ["C", "K", "Q"],
+    "answer": "C"
   },
   {
     "id": 2,
